refactor(review-page): extract internals into traits and the comment action
The page was a 2,307-line outlier, about 4x the next biggest component. It
stays one Livewire component (the 1+N hydration contract forbids nesting its
sections into children), so the extraction goes into the app layer:

- App\Concerns\ReviewPage\{ReviewsBranchDivergence, ManagesReviewTrash,
  ExportsReview} hold the cohesive clusters with little render coupling. The
  page keeps the lifecycle spine, the comment handlers, the undo() switchboard,
  the reviewed-state render glue, and every computed.
- ReviewCommentWorkflowAction gains clearAll() and restore(), so the last
  direct Comment writes leave the view file. The page applies the result via
  applyCommentMutation(), matching the existing add/update/delete path.
- CountPersistedCommentsAction owns the divergence persisted-comment read,
  keeping the cheap exists() gate alongside the count, because Livewire-layer
  traits delegate model access to Actions.
- The DTO arch allowlist recognizes App\Concerns\ReviewPage as Livewire-layer
  code: DTOs allowed, model access still Action-only.

Behavior-preserving. Pint, PHPStan, Pest, Vitest, and browser suites pass.
