+++ app/DTOs/ReviewCommentMutation.php
@@ -72,6 +72,40 @@ final readonly class ReviewCommentMutation
     }
 
     /**
+     * Every loaded comment was cleared: refresh the affected files, offer undo
+     * with the cleared rows, and re-check divergence. Renders the parent.
+     *
+     * @param  array<int, array<string, mixed>>  $comments
+     * @param  list<string>  $affectedFileIds
+     * @param  array<int, array<string, mixed>>  $clearedComments
+     */
+    public static function cleared(array $comments, array $affectedFileIds, array $clearedComments): self
+    {
+        $count = count($clearedComments);
+
+        return new self(
+            $comments,
+            $affectedFileIds,
+            ['type' => 'clear-all', 'payload' => $clearedComments, 'message' => 'Cleared '.$count.' comment'.($count === 1 ? '' : 's')],
+            checksDivergence: true,
+            skipsRender: false,
+        );
+    }
+
+    /**
+     * Removed comments were restored (the undo path): refresh the affected
+     * files and re-check divergence. Carries no undo of its own and renders the
+     * parent so the sidebar and empty states reflect the restored comments.
+     *
+     * @param  array<int, array<string, mixed>>  $comments
+     * @param  list<string>  $affectedFileIds
+     */
+    public static function restored(array $comments, array $affectedFileIds): self
+    {
+        return new self($comments, $affectedFileIds, null, checksDivergence: true, skipsRender: false);
+    }
+
+    /**
      * @return array{
      *     comments: array<int, array<string, mixed>>,
      *     affectedFileIds: list<string>,
