+++ resources/views/pages/CLAUDE.md
@@ -6,4 +6,6 @@ When designing a new page or mode, the pivotal question is **where does the user
 
 ## Known Debt
 
-- **review-page comment writes (resolved).** The comment writes (`addComment` / `addDraftComment` / `updateComment` / `deleteComment`) delegate to `ReviewCommentWorkflowAction`, which returns a `ReviewCommentMutation` DTO (new comments, affected file ids, undo payload/message, and the render-settle flags). It must be an **Action**, not a service: the arch tests forbid Livewire from using Services directly and forbid Services from touching Eloquent models, and the write path touches the `Comment` model. It is a sibling to `ContextCommentWorkflowAction`. The page keeps the dispatch and `skipRender` in `applyCommentMutation()` so the 1+N hydration contract holds (an Action cannot reach the render pipeline). **Do not** move writes into `comments-drawer`: the drawer is intentionally a lazy read-model and the writes couple to diff/session state it doesn't own. Trash/restore/clear-all stay in the page (session-state concerns, not comment concerns).
+- **review-page comment writes (resolved).** Every comment write (`addComment` / `addDraftComment` / `updateComment` / `deleteComment` / `clearAllComments` / `restoreComments`) delegates to `ReviewCommentWorkflowAction`, which returns a `ReviewCommentMutation` DTO (new comments, affected file ids, undo payload/message, and the render-settle flags) the page applies through `applyCommentMutation()`. It must be an **Action**, not a service: the arch tests forbid Livewire from using Services directly and forbid Services from touching Eloquent models, and the write path touches the `Comment` model. It is a sibling to `ContextCommentWorkflowAction`. The page keeps the dispatch and `skipRender` in `applyCommentMutation()` so the 1+N hydration contract holds (an Action cannot reach the render pipeline). **Do not** move writes into `comments-drawer`: the drawer is intentionally a lazy read-model and the writes couple to diff/session state it doesn't own.
+
+- **review-page is a coordinator over `App\Concerns\ReviewPage` traits.** The page renders N `diff-file` children, so it stays one Livewire component (nesting its sections into child components would re-hydrate every diff-file). Cohesive clusters with little render coupling are extracted into `App\Concerns\ReviewPage` traits (`ReviewsBranchDivergence`, `ManagesReviewTrash`, `ExportsReview`); the page keeps the lifecycle spine, the comment handlers, the `undo()` switchboard, the reviewed-state render glue, and every `#[Computed]`. Those traits are Livewire-layer code, so they may use DTOs but delegate all model access to Actions (e.g. `CountPersistedCommentsAction`), per the layer arch tests. The Settings dropdown and Trash list markup live in `pages/partials/` via `@include`.
